refactor: new layout and dark theme
We are now using a new responsive layout.
A layout in 2 column for xl and 2xl tailwind breakpoint and 1 column for
the all other breakpoints.

We are also implementing a dark theme version.

We are also fixing a bug in the brand page that was concatenating new
brand color cards to old ones when changin the brand color value, with
the old ones becoming all black.